requirements.txt
mysql-connector==2.2.9
pip==25.0.1
setuptools==58.0.4
paho-mqtt
